requirements.txt
aiogram==2.5 #to connect to telegram
openai==0.28.1
python-dotenv
